pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pub-sub-perf-tool"
version = "0.1.0"
description = "Generic pub-sub performance testing tool with validation support"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "kafka-python>=2.0.2",
    "pulsar-client>=3.4.0",
    "pika>=1.3.2",
    "click>=8.1.7",
    "tabulate>=0.9.0",
    "azure-eventhub>=5.11.0",
    "google-cloud-pubsub>=2.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
pub-sub-perf = "pub_sub_perf_tool.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pub_sub_perf_tool*"]
